Make PyAudio dependency optional
PyAudio is only needed for playback via the soundcard directly
from Python, not for, e.g. playing via a pipe to sox, or writing
data to a wave file.

Make it an "extras_require" depenency and guard calls to
audiogen.sampler.play() with a check to make sure PyAudio is
available.