Support multiple synchronized channels
Add a `channelize()` util function to permit synchronized multiple
channel generators.

The sampler functions all assume that if they receive a generator, it's
a single channel audio generator, and if they receive a sequence, it's a
sequence of single channel generators representing one or more channels
of audio data.

The problem is that this makes it very difficult to synchronize content
across multiple channels, since it's not possible to make a single
generator that can output samples for all channels at a particular point
in time at once.

This channelize() function lightly wraps itertools' tee() function to
make this sort of synchronized generator usable: It replicates the
multi-channel generator, one copy per output channel, then filters each
of the new generator copies so that each outputs only a single channel's
sample at each iteration.