Add variable freq, custom output range to tone()
Modify the tone() generator to accept either a fixed frequency, as
before, or a generator producing frequencies. This permits time-varying
frequncy output.

Note that the signal's phase will be continuous across varying
frequencies.

Add min_ and max_ arguments to tone() to permit output ranges other than
[-1, 1]. This is entirely equivalent to

    util.normalize(tone(frequency), -1, 1, min_, max_)

and is added solely to make tone() more convenient to use as a "pulse"
function, e.g. as an input to tone()'s own new frequency generator
argument to create a pulsed frequency:

    tone(tone(5, 440, 880))